+++ tests/DateFormatter/FormatterChainingTest.php
@@ -12,11 +12,11 @@ use Stadly\PasswordPolice\WordFormatter;
 use Traversable;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\DateFormatter\ChainableFormatter
+ * @coversDefaultClass \Stadly\PasswordPolice\DateFormatter\FormatterChaining
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class ChainableFormatterTest extends TestCase
+final class FormatterChainingTest extends TestCase
 {
     /**
      * @var MockObject&WordFormatter
@@ -34,7 +34,10 @@ final class ChainableFormatterTest extends TestCase
      */
     public function testCanSetAndGetNext(): void
     {
-        $formatter = $this->getMockForAbstractClass(ChainableFormatter::class);
+        /**
+         * @var MockObject&FormatterChaining
+         */
+        $formatter = $this->getMockForTrait(FormatterChaining::class);
 
         $formatter->setNext($this->dateFormatter);
         self::assertSame($this->dateFormatter, $formatter->getNext());
@@ -46,7 +49,10 @@ final class ChainableFormatterTest extends TestCase
      */
     public function testCanGetWhenNoNextIsSet(): void
     {
-        $formatter = $this->getMockForAbstractClass(ChainableFormatter::class);
+        /**
+         * @var MockObject&FormatterChaining
+         */
+        $formatter = $this->getMockForTrait(FormatterChaining::class);
 
         $formatter->setNext($this->dateFormatter);
         $formatter->setNext(null);
@@ -59,7 +65,10 @@ final class ChainableFormatterTest extends TestCase
      */
     public function testCanApplyFormatter(): void
     {
-        $formatter = $this->getMockForAbstractClass(ChainableFormatter::class);
+        /**
+         * @var MockObject&FormatterChaining
+         */
+        $formatter = $this->getMockForTrait(FormatterChaining::class);
         $formatter->method('applyCurrent')->willReturnCallback(
             static function (iterable $dates): Traversable {
                 foreach ($dates as $date) {
@@ -82,7 +91,10 @@ final class ChainableFormatterTest extends TestCase
      */
     public function testCanApplyFormatterChain(): void
     {
-        $formatter = $this->getMockForAbstractClass(ChainableFormatter::class);
+        /**
+         * @var MockObject&FormatterChaining
+         */
+        $formatter = $this->getMockForTrait(FormatterChaining::class);
         $formatter->method('applyCurrent')->willReturnCallback(
             static function (iterable $dates): Traversable {
                 foreach ($dates as $date) {